Remove chainToProject
