+++ packages/backend-shared/src/amounts/totalSupply.ts
