+++ packages/backend-shared/src/index.ts
@@ -1,2 +1,7 @@
-export * from './chainToProject'
-export * from './tvl'
+export * from './ConfigMapping'
+export * from './createAmountId'
+export * from './createPriceId'
+export * from './getTvlAmountsConfig'
+export * from './getTvlPricesConfig'
+export * from './indexerIdentity'
+export * from './aggLayer'
