--- packages/backend-shared/src/tvl/index.ts
@@ -1,7 +1,0 @@
-export * from './ConfigMapping'
-export * from './createAmountId'
-export * from './createPriceId'
-export * from './getTvlAmountsConfig'
-export * from './getTvlPricesConfig'
-export * from './indexerIdentity'
-export * from './aggLayer'
